+++ internal/cluster-object/object_provider.go
@@ -8,13 +8,6 @@ import (
 
 const logComponentNameKey = "component"
 
-type Manager[L client.ObjectList, O client.Object] interface {
-	ClusterScopedProvider[L]
-	NamespacedProvider[O]
-	Creator[O]
-	Deleter[L, O]
-}
-
 type Provider[L client.ObjectList, O client.Object] interface {
 	ClusterScopedProvider[L]
 	NamespacedProvider[O]
@@ -27,13 +20,4 @@ type ClusterScopedProvider[T client.ObjectList] interface {
 
 type NamespacedProvider[T client.Object] interface {
 	GetByNameAndNamespace(ctx context.Context, name, namespace string) (T, error)
-}
-
-type Creator[T client.Object] interface {
-	Create(ctx context.Context, obj T) error
-}
-
-type Deleter[L client.ObjectList, O client.Object] interface {
-	Delete(ctx context.Context, obj O) error
-	DeleteAll(ctx context.Context, objs L) error
 }
